+++ components/integrations/add_command/index.ts
@@ -2,15 +2,16 @@
 // See LICENSE.txt for license information.
 
 import {connect} from 'react-redux';
-import {bindActionCreators} from 'redux';
+import {bindActionCreators, ActionCreatorsMapObject, Dispatch} from 'redux';
 
+import {ActionFunc} from 'mattermost-redux/types/actions';
 import {addCommand} from 'mattermost-redux/actions/integrations';
 
-import AddCommand from './add_command.jsx';
+import AddCommand, {Props} from './add_command';
 
-function mapDispatchToProps(dispatch) {
+function mapDispatchToProps(dispatch: Dispatch) {
     return {
-        actions: bindActionCreators({
+        actions: bindActionCreators<ActionCreatorsMapObject<ActionFunc>, Props['actions']>({
             addCommand,
         }, dispatch),
     };
